requirements.txt
h5py~=3.9.0
matplotlib~=3.7.2
numba~=0.59.0
numpy~=1.26.4
pandas~=2.2.0
pyarrow~=15.0.2
pydantic~=2.6.0
pytest-cov~=4.1.0
pytest~=7.4.2
pyyaml~=6.0.1
scikit-learn~=1.3.0
scipy~=1.11.2
setuptools~=68.2.0
skops~=0.8.0
statsmodels~=0.14.1
tables~=3.9.2
tqdm~=4.66.1
